Fixed response code bug. Was returning a 201 sometimes even if none of the donors/samples in a bulk upload were successfully created with entity api.
Also added a new validation check to validate
sample and validate donor. In addition to making sure
every required field is there, it now also makes sure
that there are no fields present that aren't accepted.

I can think of no case where it is desireable to allow
any additional fields, but there's potential for error
If its not checked.