Added Pyyaml to requirements.txt
modified donors/bulk-upload:
* now utilizes bad_request_error if the request body is missing the file
part or the file is empty. In various parts of this method we chose to
record errors and wait until the end to return with the error message, but
in instances like this, we can't continue anyway, so it makes more sense to
end here.
* Extraneous conditionals and unnecessary boolean flags removed. Things that
were unnecessarily behind "if" conditionals no longer are.
*Condensed a number of unnecessary variable declarations to reduce lines.
* Replaced hard-coded and messy string parsing to find the bearer token with
the auth_helper getAuthorizationTokens method
* Reworked how status codes are returned in donors/bulk and samples/bulk.
Now if every entry in a tsv returns successfully, a 201 is returned. If
all of them fail, a 500 returns, and if some fail, a 207 is returned. In
each case, the result of each line in the tsv is written.

Numerous current and future bugs are removed.