Reorg of datasets/data-status and uploads/data-status has been implemented to allow caching. Uses redis.
An instance of redis server should already be running
so this uses that existing instance. Scheduler has been
created to run the logic for data-status every 5 minutes
and then updates the cache. The endpoints now only
retrieve what's on the cache. If there is an error when
reading from the cache, generates the data manually.